+++ package.js
@@ -10,7 +10,7 @@ Package.describe({
 Package.onUse(function (api) {
   api.versionsFrom('2.0')
   api.use(['meteor', 'ecmascript', 'check', 'mongo', 'typescript'])
-  api.use(['aldeed:collection2@3.3.0', 'aldeed:schema-index@3.0.0', 'aldeed:schema-deny@3.0.0'])
+  api.use(['aldeed:collection2@3.5.0', 'aldeed:schema-index@3.0.0', 'aldeed:schema-deny@3.0.0'])
   // 'socialize:base-model@1.1.2'
 
   api.mainModule('common.ts', 'client')